pyproject.toml
[project]
name = "neo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "neo4j>=5.20.0",
    "langchain>=0.2.0",
    "langchain-community>=0.2.0",
    "qdrant-client>=1.9.0",
    "langchain-qdrant>=0.1.0",
    "langchain-neo4j>=0.1.0",
    "langchain-openai>=0.1.0",
    "langchain-experimental>=0.3.4",
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "pydantic>=2.6.0",
    "python-dotenv>=1.0.0",
    "pyvis>=0.3.2",
    "pythainlp>=5.0",
]
